use params type, tool root
