+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/Root.tsx
@@ -81,6 +81,7 @@ type RootMalwareProps = {
   malwareId: string;
   queryRef: PreloadedQuery<RootMalwareQuery>;
 };
+
 const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
   const subConfig = useMemo<GraphQLSubscriptionConfig<RootMalwareSubscription>>(() => ({
     subscription,
@@ -99,11 +100,10 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
 
   const { forceUpdate, handleForceUpdate } = useForceUpdate();
 
-  const link = `/dashboard/arsenal/malwares/${malwareId}/knowledge`;
-  const isKnowledge = location.pathname.startsWith(`/dashboard/arsenal/malwares/${malwareId}/knowledge`);
   const isOverview = location.pathname === `/dashboard/arsenal/malwares/${malwareId}`;
+  const isKnowledge = location.pathname.startsWith(`/dashboard/arsenal/malwares/${malwareId}/knowledge`);
   const paddingRight = getPaddingRight(location.pathname, malwareId, '/dashboard/arsenal/malwares');
-
+  const link = `/dashboard/arsenal/malwares/${malwareId}/knowledge`;
   return (
     <>
       {malware ? (
@@ -279,9 +279,9 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
 };
 
 const Root = () => {
-  const { malwareId } = useParams();
+  const { malwareId } = useParams() as { malwareId: string; };
   const queryRef = useQueryLoading<RootMalwareQuery>(malwareQuery, {
-    id: malwareId ?? '',
+    id: malwareId,
   });
 
   return (
